move ManifestManager to :ui:base-tool
