+++ app/src/main/java/org/cru/godtools/util/ActivityUtils.kt
@@ -2,9 +2,9 @@ package org.cru.godtools.util
 
 import android.app.Activity
 import org.cru.godtools.article.activity.startCategoriesActivity
+import org.cru.godtools.base.tool.service.ManifestManager
 import org.cru.godtools.model.Tool.Type
 import org.cru.godtools.tract.activity.TractActivity
-import org.cru.godtools.xml.service.ManifestManager
 import java.util.Locale
 
 fun Activity.openToolActivity(code: String, type: Type, vararg languages: Locale) {
