+++ library/xml-model/build.gradle
@@ -1,20 +1,13 @@
 dependencies {
     implementation project(':library:analytics')
     implementation project(':library:base')
-    implementation project(':library:db')
-    implementation project(':library:model')
-    implementation project(':library:model-events')
 
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:${deps.kotlinCoroutines}"
 
-    implementation "androidx.lifecycle:lifecycle-livedata-ktx:${deps.androidX.fragment}"
-
     implementation "org.ccci.gto.android:gto-support-base:${deps.gtoSupport}"
-    implementation "org.ccci.gto.android:gto-support-compat:${deps.gtoSupport}"
     implementation "org.ccci.gto.android:gto-support-core:${deps.gtoSupport}"
     implementation "org.ccci.gto.android:gto-support-kotlin-coroutines:${deps.gtoSupport}"
-    implementation "org.ccci.gto.android:gto-support-lifecycle:${deps.gtoSupport}"
 
+    implementation "com.annimon:stream:${deps.lightweightStream}"
     implementation "com.google.guava:guava:${deps.guava}"
-    implementation "org.greenrobot:eventbus:${deps.eventbus}"
 }
